penishjakgsfdgalkjsdhalksjdgfalksdfhslkjfdasdfa

+++ backend/src/main/java/com/example/onlyonespring/repository/RoomRepository.java
@@ -3,5 +3,5 @@ package com.example.onlyonespring.repository;
 import com.example.onlyonespring.entity.FullRoom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoomRepository extends JpaRepository<FullRoom, Integer> {
+public interface RoomRepository extends JpaRepository<FullRoom, Long> {
 }
